+++ .env_example
@@ -1,3 +1,4 @@
 FLOW_PROMPT_API_TOKEN=
 AZURE_OPENAI_KEYS={"westus":{"url": "", "key": ""}}
+OPENAI_API_KEY=
 BEARER_TOKEN=
